Updated reinhard to not use internal buffer.

+++ src/raytracer/ITracer.h
@@ -16,7 +16,7 @@ public:
 	virtual void first_bounce() = 0;
 	virtual void traceImage(float* color_buffer) = 0;
 	virtual void stopTracing() = 0;
-  virtual unsigned int getPixelsLeft() = 0;
+  virtual float getPixelsLeft() = 0;
 
 };
